ARM: uniphier: Make fastbootscript as the last fallback method
Since CONFIG_PREBOOT is used for loading VOC firmware,
make fastbootscript as the last fallback method.
This invokes fastboot on USB if usb gadget can be used.
If not, it invokes fastboot on UDP.

To use this, user can run it on console, or run it at the
end of bootcmd.

Signed-off-by: Masami Hiramatsu <masami.hiramatsu@linaro.org>